+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## Misc:
 
 ## Features:
+- Added more calculation metrics to provenance and trace file to enhance reproducibility
 
 # 1.3.0
 ## Bug Fixes:
@@ -22,6 +23,7 @@
 
 ## Features:
 - Transformation of data file to provenance file with schema.org / bioschemas.org type annotation in JSON-LD data format
+- Added post-run functionality with provenance file
 - Added functionality to apply fully custom memory and CPU power functions
 
 # 1.2.1